pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestry"
version = "1.0.0"
description = "Multi-Agent Reinforcement Learning Platform for LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Orchestry Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "anthropic>=0.39.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
    "streamlit>=1.28.0",
]
tracking = [
    "wandb>=0.16.0",
]
dev = [
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pre-commit>=3.6.0",
    "types-pyyaml>=6.0.0",
    "types-tqdm>=4.66.0",
]
all = [
    "orchestry[viz,tracking,dev]",
]

[project.urls]
Homepage = "https://github.com/attaimen/Orchestry"
Repository = "https://github.com/attaimen/Orchestry"
Issues = "https://github.com/attaimen/Orchestry/issues"

[project.scripts]
orchestry-marl = "orchestry.cli.marl:main"
orchestry-legacy = "orchestry.cli.legacy:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestry*"]
exclude = ["tests*", "examples*", "docs*"]

[tool.setuptools.package-data]
orchestry = ["py.typed"]

# ============================================================================
# Black - Code Formatting
# ============================================================================
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# ============================================================================
# Ruff - Linting & Import Sorting
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "YTT",    # flake8-2020
    "ASYNC",  # flake8-async
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EXE",    # flake8-executable
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "INP",    # flake8-no-pep420
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PYI",    # flake8-pyi
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PD",     # pandas-vet
    "PGH",    # pygrep-hooks
    "PL",     # pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "NPY",    # numpy-specific
    "PERF",   # perflint
    "RUF",    # ruff-specific
]
ignore = [
    "COM812",  # trailing comma missing (conflicts with formatter)
    "ISC001",  # implicit string concatenation (conflicts with formatter)
    "T201",    # print found (we use print for CLI output)
    "TRY003",  # long messages in exceptions
    "PLR0913", # too many arguments
    "PLR2004", # magic values
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert allowed in tests
    "ARG",     # unused arguments OK in tests
    "PLR2004", # magic values OK in tests
]
"examples/**/*.py" = [
    "T201",    # print OK in examples
    "INP001",  # no __init__ needed in examples
]

[tool.ruff.lint.isort]
known-first-party = ["orchestry"]
force-sort-within-sections = true
lines-after-imports = 2

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-args = 8
max-branches = 12

# ============================================================================
# Mypy - Type Checking (Strict Mode)
# ============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Third-party packages without type stubs or with incomplete stubs
[[tool.mypy.overrides]]
module = [
    "anthropic.*",
    "matplotlib.*",
    "plotly.*",
    "streamlit.*",
    "wandb.*",
    "dotenv",
    "rich.*",
    "tqdm",
    "numpy.*",
]
ignore_missing_imports = true

# Legacy code - skip strict checking entirely
[[tool.mypy.overrides]]
module = [
    "orchestry.legacy.*",
    "orchestry.cli.legacy",
]
# Legacy code kept for compatibility - not actively maintained
follow_imports = "skip"
ignore_errors = true

# ============================================================================
# Pytest - Testing
# ============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=orchestry",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["orchestry"]
omit = [
    "*/tests/*",
    "*/examples/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false
